Update funside.mdx
---
title: Funtoo Linux on the RPi400 (OS Review)

description: Funtoo Linux + a raspberry pi, what could go wrong?
date: June 13 2022
---

# So I installed funtoo on my Raspberry Pi.

Some of you readers out there might have a small inkling as to what I'm talking about here. For you: it's a better experience than on an AMD Jaguar chip.
![Screenshot_2022-06-09_10-03-15](https://user-images.githubusercontent.com/19198652/173349754-41086da1-86ca-45ad-a074-c4c95e963ce9.png)
I installed it a couple of weeks ago now. It's a far faster os once every thing is compiled than even ARCH is on the pi, and it's easier to get to that state than Gentoo is. Ego is such a great tool to work with and their version of portage is pretty fast.

I had a setup that made XFCE4 look like Unity7, hence the screenshot I shared. It's not a thing exclusive to funtoo, but funtoo definitely makes it fun to do. I prefer plasma or mate for the task of getting a Unity7 lookalike, however. I don't think that Gentoo and Funtoo are fairly comparable despite using the same package manager. The focuses and directions of the two projects are nowhere near the same and Funtoo has a focused build for the raspberry pi. Funtoo Linux is a fun project to look at on your computer or raspberry pi.